Split out different styles into own component
