+++ blocks/bauhaus/src/colors.js
@@ -1,0 +1,12 @@
+/**
+ * WordPress dependencies
+ */
+import { __ } from '@wordpress/i18n';
+
+export default [
+	{ name: __( 'Black' ), color: '#000000' },
+	{ name: __( 'Blue' ), color: '#051BF4' },
+	{ name: __( 'Red' ), color: '#D32121' },
+	{ name: __( 'Yellow' ), color: '#F7FC1C' },
+	{ name: __( 'White' ), color: '#FFFFFF' },
+];
